(fix) GatewayHttpClient not resolving base url correctly
