+++ hummingbot/client/command/gateway_command.py
@@ -36,7 +36,7 @@ from hummingbot.core.utils.gateway_config_utils import (
     native_tokens,
     upsert_connection
 )
-from hummingbot.core.gateway import gateway_http_client
+from hummingbot.core.gateway.gateway_http_client import GatewayHttpClient
 from hummingbot.core.utils.ssl_cert import certs_files_exist, create_self_sign_certs
 from hummingbot.client.config.config_helpers import save_to_yml
 from hummingbot.client.config.global_config_map import global_config_map
@@ -131,7 +131,7 @@ class GatewayCommand:
 
     async def _test_connection(self):
         # test that the gateway is running
-        if await gateway_http_client.ping_gateway():
+        if await GatewayHttpClient.get_instance().ping_gateway():
             self.notify("\nSuccesfully pinged gateway.")
         else:
             self.notify("\nUnable to ping gateway.")
@@ -162,7 +162,7 @@ class GatewayCommand:
         self.app.hide_input = False
         self.app.change_prompt(prompt=">>> ")
 
-        gateway_http_client.reload_certs()
+        GatewayHttpClient.get_instance().reload_certs()
 
     async def _generate_gateway_confs(self, container_id: str, conf_path: str = "/usr/src/app/conf"):
         self.app.clear_input()
@@ -307,7 +307,7 @@ class GatewayCommand:
 
         if self._gateway_monitor.current_status == Status.ONLINE:
             try:
-                status = await gateway_http_client.get_gateway_status()
+                status = await GatewayHttpClient.get_instance().get_gateway_status()
                 self.notify(pd.DataFrame(status))
             except Exception:
                 self.notify("\nError: Unable to fetch status of connected Gateway server.")
@@ -316,7 +316,7 @@ class GatewayCommand:
 
     async def _update_gateway_configuration(self, key: str, value: Any):
         try:
-            response = await gateway_http_client.update_config(key, value)
+            response = await GatewayHttpClient.get_instance().update_config(key, value)
             self.notify(response["message"])
         except Exception:
             self.notify("\nError: Gateway configuration update failed. See log file for more details.")
@@ -357,7 +357,7 @@ class GatewayCommand:
 
         else:
             # get available networks
-            connector_configs = await gateway_http_client.get_connectors()
+            connector_configs = await GatewayHttpClient.get_instance().get_connectors()
             connector_config = [d for d in connector_configs["connectors"] if d["name"] == connector]
             available_networks = connector_config[0]["available_networks"]
             trading_type = connector_config[0]["trading_type"][0]
@@ -388,7 +388,7 @@ class GatewayCommand:
                     self.notify("Error: Invalid network")
 
             # get wallets for the selected chain
-            wallets_response = await gateway_http_client.get_wallets()
+            wallets_response = await GatewayHttpClient.get_instance().get_wallets()
             wallets = [w for w in wallets_response if w["chain"] == chain]
             if len(wallets) < 1:
                 wallets = []
@@ -400,7 +400,7 @@ class GatewayCommand:
                 self.app.clear_input()
                 self.placeholder_mode = True
                 wallet_private_key = await self.app.prompt(prompt=f"Enter your {chain}-{network} wallet private key >>> ")
-                response: Dict[str, Any] = await gateway_http_client.add_wallet(chain, network, wallet_private_key)
+                response: Dict[str, Any] = await GatewayHttpClient.get_instance().add_wallet(chain, network, wallet_private_key)
                 wallet_address: str = response["address"]
 
             # the user has a wallet. Ask if they want to use it or create a new one.
@@ -416,7 +416,7 @@ class GatewayCommand:
                     native_token = native_tokens[chain]
                     wallet_table = []
                     for w in wallets:
-                        balances: Dict[str, Any] = await gateway_http_client.get_balances(
+                        balances: Dict[str, Any] = await GatewayHttpClient.get_instance().get_balances(
                             chain, network, w, [native_token]
                         )
                         wallet_table.append({"balance": balances['balances'][native_token], "address": w})
@@ -438,7 +438,7 @@ class GatewayCommand:
                 else:
                     self.placeholder_mode = True
                     wallet_private_key = await self.app.prompt(prompt=f"Enter your {chain}-{network} wallet private key >>> ")
-                    response = await gateway_http_client.add_wallet(chain, network, wallet_private_key)
+                    response = await GatewayHttpClient.get_instance().add_wallet(chain, network, wallet_private_key)
                     wallet_address = response["address"]
 
             # write wallets to json
@@ -458,7 +458,7 @@ class GatewayCommand:
 
     @staticmethod
     async def _fetch_gateway_configs() -> Dict[str, Any]:
-        return await gateway_http_client.get_configuration(fail_silently=True)
+        return await GatewayHttpClient.get_instance().get_configuration(fail_silently=True)
 
     async def fetch_gateway_config_key_list(self):
         config = await self._fetch_gateway_configs()
