+++ hummingbot/strategy/amm_arb/utils.py
@@ -40,5 +40,6 @@ async def create_arb_proposals(market_info_1: MarketTradingPairTuple,
             m_2_o_price,
             order_amount
         )
+
         results.append(ArbProposal(first_side, second_side))
     return results
